requirements.txt
dotmap==1.3.23
future==0.18.2
gpflow==1.5.1
gym==0.17.3
matplotlib==3.3.4
mujoco-py==1.50.1.68
numpy==1.19.5
scipy==1.5.4
termcolor==1.1.0
tqdm==4.60.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
protobuf==3.13.0
